Cip branch

+++ src/main/java/ryleh/model/components/LurkerComponent.java
@@ -29,7 +29,7 @@ public class LurkerComponent extends Component {
 	    }
 
 	    @Override
-	    public void onUpdate(final int deltaTime) {
+	    public void onUpdate(final double deltaTime) {
 	        move();
 	       // rotate();
 	    }
