+++ src/main/java/ryleh/view/enemies/EnemyShooterGraphicComponent.java
@@ -24,7 +24,7 @@ public class EnemyShooterGraphicComponent implements GraphicComponent{
 	}
 
 	@Override
-	public void render(final Point2D position, final int deltaTime) {
+	public void render(final Point2D position, final double deltaTime) {
 		rectangle.setX(position.getX());
 		rectangle.setY(position.getY());
 		this.updateImage();
