+++ src/main/java/ryleh/view/other/RockGraphicComponent.java
@@ -22,7 +22,7 @@ public class RockGraphicComponent implements GraphicComponent{
 	}
 
 	@Override
-	public void render(final Point2D position, final int deltaTime) {
+	public void render(final Point2D position, final double deltaTime) {
 		rectangle.setX(position.getX());
 		rectangle.setY(position.getY());
 	}
